jsp_thmyleaf/requirements.txt
openai~=0.28.1
requests==2.31.0
PyYAML~=6.0
from-root